QueryDSL을 이용하여 장치 통계 및 그룹 통계 기능 구현

+++ src/main/java/com/namji/datacollection/dto/response/StatisticsResponse.java
@@ -9,6 +9,6 @@ public class StatisticsResponse {
 
   private Long deviceId;
   private String serialNumber;
-  private float averageData;
+  private Double averageData;
 
 }
